+++ b4n/core/app.rs
@@ -260,6 +260,7 @@ impl App {
             ResponseEvent::ShowPortForwards => self.views_manager.show_port_forwards(),
             ResponseEvent::PortForward(resource, to, from, address) => self.port_forward(resource, to, from, &address),
             ResponseEvent::RunPlugin(id, context) => self.views_manager.run_plugin(&id, context),
+            ResponseEvent::InjectContainer(resource, container) => self.views_manager.inject_container(&resource, container),
             _ => (),
         }
 
@@ -280,6 +281,7 @@ impl App {
                 CommandResult::ThemesList(list) => self.views_manager.show_themes_list(list),
                 CommandResult::ResourcePortsList(list) => self.views_manager.show_ports_list(&list),
                 CommandResult::RunPluginOutput(result) => self.views_manager.show_plugin_output(&command.id, result),
+                CommandResult::InjectedContainer(result) => self.views_manager.show_inject_result(result),
             }
         }
     }