+++ b4n/core/managers/views.rs
@@ -1,15 +1,15 @@
 use anyhow::Result;
-use b4n_common::{DEFAULT_ERROR_DURATION, IconKind, NotificationSink};
+use b4n_common::{DEFAULT_ERROR_DURATION, DEFAULT_MESSAGE_DURATION, IconKind, NotificationSink};
 use b4n_config::keys::KeyCommand;
 use b4n_kube::plugins::PluginContext;
 use b4n_kube::{
     ALL_NAMESPACES, ContainerRef, Namespace, PODS, Port, PropagationPolicy, ResourceRef, ResourceRefFilter, ResourceTag,
 };
 use b4n_tasks::commands::{
-    CommandResult, DeleteResourcesOptions, GetNewResourceYamlError, GetNewResourceYamlResult, ResourceYamlError,
-    ResourceYamlResult, RunPluginError, RunPluginOutput, SetNewResourceYamlError, SetResourceYamlError,
+    CommandResult, DeleteResourcesOptions, GetNewResourceYamlError, GetNewResourceYamlResult, InjectContainerError,
+    ResourceYamlError, ResourceYamlResult, RunPluginError, RunPluginOutput, SetNewResourceYamlError, SetResourceYamlError,
 };
-use b4n_tui::{MouseEventKind, ResponseEvent, Responsive, ToSelectData, TuiEvent};
+use b4n_tui::{EphemeralContainer, MouseEventKind, ResponseEvent, Responsive, ToSelectData, TuiEvent};
 use b4n_tui::{table::Table, table::ViewType, widgets::Footer};
 use kube::{config::NamedContext, discovery::Scope};
 use ratatui::layout::{Constraint, Direction, Layout, Rect};
@@ -644,6 +644,36 @@ impl ViewsManager {
             self.view = None;
         } else if let Some(view) = &mut self.view {
             view.process_command_result(CommandResult::RunPluginOutput(result));
+        }
+    }
+
+    /// Runs command to inject ephemeral container to the specified resource.
+    pub fn inject_container(&mut self, resource: &ResourceRef, container: EphemeralContainer) {
+        self.worker.borrow_mut().inject_container(resource, container);
+        self.footer().show_info(
+            format!(
+                "Injecting ephemeral container '{}' into '{}' pod",
+                resource.container.as_deref().unwrap_or_default(),
+                resource.name.as_deref().unwrap_or_default()
+            ),
+            DEFAULT_MESSAGE_DURATION,
+        );
+    }
+
+    /// Shows result from the ephemeral container injection in the footer.
+    pub fn show_inject_result(&mut self, result: Result<ResourceRef, InjectContainerError>) {
+        match result {
+            Ok(resource) => self.footer().show_info(
+                format!(
+                    "Ephemeral container '{}' successfully injected into '{}' pod",
+                    resource.container.as_deref().unwrap_or_default(),
+                    resource.name.as_deref().unwrap_or_default()
+                ),
+                DEFAULT_MESSAGE_DURATION,
+            ),
+            Err(error) => self
+                .footer()
+                .show_error(format!("Ephemeral container error: {}", error), DEFAULT_ERROR_DURATION),
         }
     }