+++ b4n/ui/views/resources/mod.rs
@@ -1,5 +1,6 @@
 pub use self::table::*;
 pub use self::view::*;
 
+mod menus;
 mod table;
 mod view;