+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- ephemeral containers support
 - highlight command palette accept button on mouse move
 
 ### Bug fixes